requirements.txt
customtkinter
winotify
groq
python-dotenv
requests
